fix: runner result adapter
